lp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
huggingface-hub==0.25.0
cryptography==43.0.1
python-multipart==0.0.12
